requirements.txt
discord.py
python-dotenv
bs4
newsapi-python
numpy
tensorflow
pandas
scikit-learn
matplotlib
alpha-vantage
